Another take on inferred joining.
I think I like this one a bit better -- it defers changing where
values to predicates until build_arel is called, and is turned on
for a relation by putting "autojoin" in the chain.

Also, it does away with the mw_infer and all the craziness with
modifying the where condition hashes on the fly.